updated linktaflow schema and id mapping logic
