+++ server/types/index.ts
@@ -1,2 +1,4 @@
 export * from './errorHandling.d';
-export * from './datamodels.d';
+export * from './userInput';
+export * from './linktaFlow';
+export * from './user';
