+++ server/types/userInput.ts
@@ -1,0 +1,12 @@
+import type { Types } from 'mongoose';
+
+export type UserInput = {
+  _id?: Types.ObjectId;
+  id: Types.ObjectId;
+  userId: Types.ObjectId;
+  title: string;
+  input: string;
+  linktaFlowId: Types.ObjectId;
+  createdAt: Date;
+  updatedAt: Date;
+};
